cmd/contour: move serve subcommand handling to doServe
